# PERF: reducing dtype checking overhead in groupby