@jreback 2021-12-03
yeah looks nice, i agree should try to refactor is_* to make them more strict is a good idea