@jbrockmendel 2021-12-03
The dtype.kind checks i definitely like.  How big a difference does the caching make?