@jreback 2022-02-17
this would be nice to do